fix: MapProvider hooks 분리
